app logic moved to api.py
